更新 代码jdk17优化

+++ reed-common/src/main/java/top/reed/common/utils/ShiroUtils.java
@@ -47,7 +47,7 @@ public class ShiroUtils {
 	}
 
 	public static Long getUserId() {
-		return getSysUser().getUserId().longValue();
+		return getSysUser().getUserId();
 	}
 
 	public static String getLoginName() {
