+++ reed-framework/src/main/java/top/reed/framework/shiro/service/SysPasswordService.java
@@ -30,7 +30,7 @@ public class SysPasswordService {
 
 	private Cache<String, AtomicInteger> loginRecordCache;
 
-	@Value(value = "${user.password.maxRetryCount}" )
+	@Value(value = "${user.password.maxRetryCount}")
 	private String maxRetryCount;
 
 	@PostConstruct
@@ -47,13 +47,13 @@ public class SysPasswordService {
 			retryCount = new AtomicInteger(0);
 			loginRecordCache.put(loginName, retryCount);
 		}
-		if (retryCount.incrementAndGet() > Integer.valueOf(maxRetryCount).intValue()) {
-			AsyncManager.me().execute(AsyncFactory.recordLogininfor(loginName, Constants.LOGIN_FAIL, MessageUtils.message("user.password.retry.limit.exceed" , maxRetryCount)));
-			throw new UserPasswordRetryLimitExceedException(Integer.valueOf(maxRetryCount).intValue());
+		if (retryCount.incrementAndGet() > Integer.valueOf(maxRetryCount)) {
+			AsyncManager.me().execute(AsyncFactory.recordLogininfor(loginName, Constants.LOGIN_FAIL, MessageUtils.message("user.password.retry.limit.exceed", maxRetryCount)));
+			throw new UserPasswordRetryLimitExceedException(Integer.valueOf(maxRetryCount));
 		}
 
 		if (!matches(user, password)) {
-			AsyncManager.me().execute(AsyncFactory.recordLogininfor(loginName, Constants.LOGIN_FAIL, MessageUtils.message("user.password.retry.limit.count" , retryCount)));
+			AsyncManager.me().execute(AsyncFactory.recordLogininfor(loginName, Constants.LOGIN_FAIL, MessageUtils.message("user.password.retry.limit.count", retryCount)));
 			loginRecordCache.put(loginName, retryCount);
 			throw new UserPasswordNotMatchException();
 		} else {
